Fixed issue #8
Turns out that `projmake-elmm/remove-project-list-buffer` was being
passed a `project`, and then trying to pull a project out of it as if it
were a `build-state`. I'm not quite sure how this was working for
anyone.

Also, I closed the window of the error-list buffer, only if there were
no errors.