Fix problem where fringe does not exist in terminal emacs
At times the 'indicators' library doesn't exist because fringe doesn't
exist in terminal emacs. This allows projmake-mode to continue to
function even when this is the case.

changes. Lines starting # with '#' will be ignored, and an empty
message aborts the commit.  # On branch master # Your branch is
up-to-date with 'origin/master'.  # # Changes to be committed: #
modified: projmake-markup.el #